Moved unused files to a dir called "private"
